timer: fix multiple processing of timer lists
The statement in the timer process function:
/* If the next entry is removed by the callback here,
 * the list has to be processed again from the begining as
 * 'n' will point to the deleted entry's next value.
 */

is wrong as doing so will break 'remaining_loops' of the timers
that have already been processed in the list!
We have to remove all entries from the list and reschedule
them if needed.

The current solution will always remove timers from the processed list
and reschedule them to the (current index - 1) position.
This way we can also get rid of timer statuses and rely on the timer list
pointer wich will indicate if a given timer is scheduled or not.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>