scheduler: reduce the risk of having irqs disabled forever
A task can be rescheduled until its job is done. If interrupts are disabled,
because the task water mark has been reached, a task may never finish its job
if it needs an interrupt handler to complete it.

This typical situation can happen if there are no more free packet buffers
and only an interrupt handler can free them but this won't happen as interrupts
are disabled. A task that asks for a packet buffer will reschedule itself until
packets are available again which will never happen.

To fix this behavior, we must re-enable interrupts as soon as possible and not
wait until all tasks have been processed by the scheduler.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>