apps/alarm: be able to register new devices
This patch allows to register new devices on address 0xFE.
The new devices must advertise their features which will be stored
in the EEPROM. A first available address will be assigned to a new
device and a new connection will be assiciated to that address.

In order to prevend the master module to initiate new swen associations,
the swen_l3_associate() function is removed from the retransmission
handler. The association must be handled in the application.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>