net/swen-l3: notify application when peer re-connects
Currently, when the peer re-connects without closing an existing
association (this can happen after a reboot or a crash), the callback
function handling the association events is not notified.
This is not good as the application can miss important events.

For example when the peer sends a command and reboots for whatever reason
before the command is actually sent, the peer will never get the command
and will know nothing about the end-point re-connect.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>