net/swen-l3: fix handling reserved bytes
Two bytes are reserved when sending a swen-l3 packet:
one for the 'retries left' number and one for the sequence ID.

These bytes are wrongly reserved at the end of the packet data
because the check for the packet free space does not take into
account the encryption with xtea which can add additional 4 bytes
of payload.
To handle the packet length safely these bytes should be reserved
at the beginning of the packet.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>