scheduler: do run scheduled tasks from within the schedule_task() function
We cannot garantee if the task ring is in the right state when
executing the scheduler from an interrupt handler.
Some tasks may be executed twice if the scheduler is interrupted and
ran again from the interrupt handler.
Also do not abort if a task cannot be scheduled, just skip it.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>