drivers/rf: fix packet starvation on receiving
In order to refill the interface packet pool receiving process has to be
started. But this will lead to a race condition if the packet pool is already
empty and the reception has not begun yet.
This patch fixes that by refilling the interface packet pool in case the packet
allocation from it fails.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>