drivers/rf: add delays between subsequent sends
The pulse of the RF433Mhz devices is not of the same width at the beginning
of the transmission and at its end. The variations in the pulse width make
long (> 15 bytes) transmissions fail.
To overcome this, a delay between each packet to send is added so the
transmitter starts sending every packet with the same pulse width.
This is not ideal as long packets will encounter the same issue but
it fixes efficiently sending multiple small packets.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>